requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.0
opencv-python==3.4.5.20
opt-einsum==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
